extend the gym display with openable member details
When a gym member (pokémon) is clicked, it's information is shown. This helps mobile users as the trainer name tooltips were nearly impossible to see on touch.

The issue required some reorganization of the gym details layout, nothing ground-breaking. The pokemon icons now include a fixed size border which is transparent by default but modifies the spacing of icons.

Information includes CP, IVs, HP and moves, so it lets you prepare for a gym battle. This required new fields in the raw_data response.

Move names were not available so I made a json from all the moves listed on bulbapedia. The fields are self explanatory I hope.

The team logo is supposed to be the same size as the details texts. This is true in chrome, but not perfect in firefox or edge. Might need some tuning, a fixed height container or the whole idea should just be scrapped as the team icon holds no valuable information whatsoever.

Stuff is in english, translations are completely neglected, sorry.